feat(std): carry a 64-bit int across the standard library's C boundary
The three shipped C sources declared candela's int as int32_t, which is
now half the width the caller passes and returns. They take and return
int64_t instead: time::now outlives 2038, math clamps an exponent to what
C's ldexp and scalbn accept, and random draws and bounds run at the full
width, two 32-bit draws to a value with the same rejection sampling the
32-bit bounded draw used.

+++ libs/std_src/time/time.c
@@ -7,8 +7,8 @@
 #include <stdint.h>
 #include <time.h>
 
-EXPORT int32_t now(void) { return (int32_t)time(NULL); }
-EXPORT const char *format(int32_t timestamp, const char *fmt) {
+EXPORT int64_t now(void) { return (int64_t)time(NULL); }
+EXPORT const char *format(int64_t timestamp, const char *fmt) {
   static char buffer[128];
   time_t t = (time_t)timestamp;
   struct tm *info = localtime(&t);
